% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{advance}
\alias{advance}
\title{Advance age classes.}
\usage{
advance(N, R = 0, plus = FALSE)
}
\arguments{
\item{N}{projected abundance}

\item{R}{number of recruits}

\item{plus}{whether the final age class is a plus class}
}
\value{
A vector of initial abundances for the next year.
}
\description{
Given projected abundances, construct the initial abundances for
the next projection.
}
\details{
At the end of each projection, the final cohort numbers advance by
one age class to become the initial numbers for the next
projection. If the final class is a "plus" class, the numbers from
the previous year are retained, otherwise the final age class is
discarded.
}
\seealso{
\code{\link[=project]{project()}}, \code{\link[=projectC]{projectC()}} for annual projection
}
