% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date.r
\name{check_dates}
\alias{check_dates}
\title{Checking increment dates}
\usage{
check_dates(inc_1 = "2019-01-01", inc_2 = 1)
}
\arguments{
\item{inc_1}{The start date of the projection year in the ISO8601 format "YYYY-MM-DD"}

\item{inc_2}{Either an increment number or a date in the ISO8601 format "YYYY-MM-DD"}
}
\value{
When inc_2 is an increment number \code{check_dates} returns the date that increment refers to. If given a date \code{check_dates} returns the increment number in the year from the start date.
}
\description{
\lifecycle{experimental}
In order to check the dates when using 365 increments in a Grym simulation this function helps translate the dates into increments or the other way around. The function will warn you if either of the dates fall in a leap year and the increment number or date may be out by 1. If this occurs, recommend changing the years to get correct increment numbers.
}
\examples{
check_dates(inc_1="2018-12-01", inc_2="2019-01-21")
check_dates(inc_1="2018-12-01", inc_2=94)

}
