% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{doubleNormalSelectivity}
\alias{doubleNormalSelectivity}
\title{Double Normal Selectivity}
\usage{
doubleNormalSelectivity(x, x1, x2, s1, s2, y1, y2)
}
\arguments{
\item{x}{a matrix or vector}

\item{x1}{endpoint of the ascending limb}

\item{x2}{endpoint of the descending limb}

\item{s1}{width of the ascending limb}

\item{s2}{width of the descending limb}

\item{y1}{minimum selectivity the ascending limb}

\item{y2}{dispersion of the descending limb}
}
\value{
A matrix or vector of probabilities of the same size as x
}
\description{
Double Normal Selectivity.
}
\details{
Compute a double Normal or "dome shaped" selectivity function, parameterized in terms of the
endpoint, width and minima of the ascending and descending limbs.
}
\examples{
x <- seq(from=150, to=450, by=2)
p <- doubleNormalSelectivity(x, x1=250, x2=320, s1=30, s2=15, y1=0.1, y2=0.7)
plot(x=x, y=p, type="l")
}
