% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Grym-package}
\alias{Grym-package}
\title{R Implementation of the Generalized Yield Model}
\description{
Grym is an R Implementation of the Generalized Yield Model (GYM) - a
flexible model for fish stock projection.
}
\author{
S. Wotherspoon, D. Maschette
}
