% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{project}
\alias{project}
\title{Yearly projections for known fishing mortality.}
\usage{
project(
  ws,
  MMs,
  FFs = 0,
  Ffs = 0,
  Nref = 1,
  nref = 1,
  Bref = NA,
  bref = nref,
  yield = 2
)
}
\arguments{
\item{ws}{matrix of weight at age}

\item{MMs}{matrix of \emph{scaled} integrated natural mortalities}

\item{FFs}{matrix of \emph{scaled} integrated fishing mortalities}

\item{Ffs}{matrix of \emph{scaled} fishing mortalities}

\item{Nref}{vector of reference abundances for each age class}

\item{nref}{vector of time steps at which the reference abundance
is estimated}

\item{Bref}{optional reference biomass}

\item{bref}{vector of time steps at which the reference biomass is
estimated}

\item{yield}{should yield be calculated for the final time step,
ever time step, or not at all (see details)}
}
\value{
\code{project} returns a list with elements
\itemize{
\item \code{N} matrix of abundance
\item \code{B} matrix of biomass
\item \code{Y} matrix or vector of yield
}
}
\description{
Project the abundance, biomass and yield in each age class forward
over one year when natural and fishing mortalities are known.
}
\details{
The \code{project} function calculates the annual projection of
abundance \code{N}, biomass \code{B} and yield \code{Y} that reproduces the
prescribed reference abundance or biomass given known natural and
fishing mortalities.  This function requires the natural and
fishing mortality be prescribed in the form of the \emph{scaled}
integrated natural and fishing mortalities and the \emph{scaled}
instantaneous fishing mortality.

When no reference biomass is specified, the vector \code{Nref} of
reference abundances specifies the mean abundance of each age
class for the time steps specified by \code{nref} (\code{nref=1} corresponds
to specifying an initial abundance).  When a reference biomass
\code{Bref} is specified, \code{Nref} is interpreted as relative abundances
and the projections are scaled to the reference biomass.

The matrices \code{ws}, \code{MMs}, \code{FFs}, \code{Ffs}, \code{N}, \code{B} and \code{F} are all
of equal size, and have one column for each age class, and one row
for each time step through the year.  So if a daily time step is
adopted, each matrix will have 366 rows with the first
representing the midnight on the last day of the previous year,
and the last representing midnight on the last day of the current
year.

If \code{yield=0}, then yield is not calculated.  If \code{yield=1}, then
only the final yield from each age class is calculated and \code{Y} is
a vector with one element for each age class.  If \code{yield=2} then
\code{Y} is the yield per time step and age class and is a matrix the
same size as \code{N} and \code{B}.
}
\examples{
## Daily time steps and 7 age classes
nsteps <- 365
Ages <- 2:8
Days <- seq(from=0, to=1, length=nsteps+1)
h <- 1/nsteps

## Ages
ages <- outer(X=Days, Y=Ages, FUN="+")
## Age-length and length-weight conversions
ls <- vonBertalanffyAL(A=ages, t0=0.0667, K=0.5, Linf=500)
ws <- powerLW(L=ls, a=9E-10, b=3.32)

## Constant intra-annual natural mortality
ms <- matrix(data=1, nrow=nsteps+1, ncol=length(x=Ages))
ms <- ms/mean(x=trapz(fs=ms, h=h))
Ms <- ctrapz(fs=ms, h=h)
Msf <- final(P=Ms)
M <- 0.1

## Normalized within year distribution of fishing effort
fwy <- double(length=nsteps+1)
fwy[81:220] <- 1
fwy <- fwy/trapz(fs=fwy, h=h)
## Length based selectivity
ss <- rampOgive(x=ls, x50=420, xrange=30)

## Intra-annual fishing mortality - effort by selectivity
fs <- fwy*ss
Fs <- ctrapz(fs=fs, h=h)
Fsf <- final(P=Fs)
F <- 0.2

## Deterministic initial age structure
N0 <- ageStructureD(MMsf=M*Msf, FFsf=F*Fsf)
N0

## Annual projection from N0
pr <- project(ws=ws, MMs=M*Ms, FFs=F*Fs, Ffs=F*fs, Nref=N0, yield=2)

## Final abundance
pr$N[nsteps+1,]
## Total annual yield
sum(pr$Y[nsteps+1,])

## Plot initial abundance, projected abundance, biomass and yield
opar <- par(mfrow=c(2,2), mar=c(5,4,2,2)+0.1)
pal <- hcl(h=seq(from=15, to=375, length=8)[1:7], l=65, c=100)
plot(x=Ages, y=N0, xlab="Age", ylab="Abundance")
matplot(x=Days, y=pr$N, xlab="Day", ylab="Abundance", type="l", lty=1, col=pal)
matplot(x=Days, y=pr$B, xlab="Day", ylab="Biomass", type="l", lty=1, col=pal)
matplot(x=Days, y=pr$Y, xlab="Day", ylab="Yield", type="l", lty=1, col=pal)
par(opar)

## Annual projection given relative abundance and total
## biomass estimated from a five day mid-year survey (days 150 to 154)
Nsurvey <- c(1.0,0.67,0.46,0.33,0.25,0.18,0.13) 
Bsurvey <- 15000 
pr <- project(ws=ws, MMs=M*Ms, FFs=F*Fs, Ffs=F*fs, Nref=Nsurvey, nref=151:155, Bref=Bsurvey, yield=2)

## Total annual yield
sum(pr$Y[nsteps+1,])

## Plot initial abundance, projected abundance, biomass and yield
opar <- par(mfrow=c(2,2), mar=c(5,4,2,2)+0.1)
pal <- hcl(h=seq(from=15, to=375, length=8)[1:7],l=65,c=100)
plot(x=Ages, y=pr$N[1,], xlab="Age", ylab="Abundance")
matplot(x=Days, y=pr$N, xlab="Day", ylab="Abundance", type="l" ,lty=1, col=pal)
matplot(x=Days, y=pr$B, xlab="Day", ylab="Biomass", type="l", lty=1, col=pal)
matplot(x=Days, y=pr$Y, xlab="Day", ylab="Yield", type="l", lty=1, col=pal)
par(opar)

## Ten year projection from virgin stock collating final annual
## abundance, biomass and yield
Years <- 1:10
Nf <- Bf <- Yf <- matrix(data=NA, nrow=length(Years), ncol= length(Ages))

## Deterministic initial age structure assuming no fishing and 8000
## new recruits each year
N0 <- ageStructureD(MMsf=M*Msf, R=8000)

## Project forward with constant fishing mortality
for(k in 1:nrow(x=Nf)) {
  pr <- project(ws=ws, MMs=M*Ms, FFs=F*Fs, Ffs=F*fs, Nref=N0, yield=1)
  N0 <- advance(N=pr$N, R=8000)
  Nf[k,] <- final(P=pr$N)
  Bf[k,] <- final(P=pr$B)
  Yf[k,] <- pr$Y
}

## Plot annual abundance, biomass and yield
opar <- par(mfrow=c(2,2), mar=c(5,4,2,2)+0.1)
pal <- hcl(h=seq(from=15, to=375, length=8)[1:7], l=65, c=100)
matplot(x=Years, y=Nf, xlab="Year", ylab="Abundance", type="l", lty=1, col=pal)
matplot(x=Years, y=Bf, xlab="Year", ylab="Biomass", type="l", lty=1, col=pal)
matplot(x=Years, y=Yf, xlab="Year", ylab="Yield", type="l", lty=1, col=pal)
par(opar) 
}
\seealso{
\code{\link[=projectC]{projectC()}}, \code{\link[=advance]{advance()}} for annual projection,
\code{\link[=ageStructureD]{ageStructureD()}}, \code{\link[=ageStructureS]{ageStructureS()}} for initial age structure
}
