% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{projectC}
\alias{projectC}
\title{Yearly projections for known catch.}
\usage{
projectC(
  ws,
  MMs,
  Fs,
  fs,
  Catch,
  Nref,
  nref = 1,
  Bref = NA,
  bref = nref,
  yield = 2,
  Fmax = 2.5,
  tol = 1e-06
)
}
\arguments{
\item{ws}{matrix of weight at age}

\item{MMs}{matrix of \emph{scaled} integrated natural mortalities}

\item{Fs}{matrix of \emph{unscaled} integrated fishing mortalities}

\item{fs}{matrix of \emph{unscaled} fishing mortalities}

\item{Catch}{target total annual catch}

\item{Nref}{vector of reference abundances for each age class}

\item{nref}{vector of time steps at which the reference abundance is estimated}

\item{Bref}{optional reference biomass}

\item{bref}{vector of time steps at which the reference biomass is estimated}

\item{yield}{should yield be calculated for the final time step,
ever time step, or not at all (see details)}

\item{Fmax}{the maximum reasonable annual scaling of fishing mortality}

\item{tol}{the tolerance on target catch}
}
\value{
\code{projectC} returns a list with elements
\itemize{
\item \code{N} matrix of abundance
\item \code{B} matrix of biomass
\item \code{Y} vector or matrix of yield
\item \code{F} annual fishing mortality required to reproduce the observed catch.
}
}
\description{
Projection of abundance, biomass and yield by age class forward
over a year when natural mortality and total annual catch are known.
}
\details{
The \code{projectC} function calculates the annual projection of
abundance \code{N}, biomass \code{B} and yield \code{Y} that reproduces the
prescribed reference abundance or biomass given known natural
mortality and total catch. This function requires the \emph{scaled}
integrated natural mortality, and the \emph{unscaled} integrated and
instantaneous fishing mortalities, and it repeatedly calls
\code{project} to determine the annual scaling of fishing mortality \code{F}
that reproduces the target catch.

If the target catch cannot be recovered with an \code{F} smaller than
\code{Fmax}, then the projection is performed with \code{F} capped at
\code{Fmax}, and it is the responsibility of the user to check if the
target catch has been recovered.  As the target catch is only ever
approximately recovered it is advised to test for \code{F==Fmax}.

When no reference biomass is specified, the vector \code{Nref} of
reference abundances specifies the mean abundance of each age
class for the time steps specified by \code{nref} (\code{nref=1} corresponds
to specifying an initial abundance).  When a reference biomass
\code{Bref} is specified, \code{Nref} is interpreted as relative abundances
and the projections are scaled to the reference biomass.

The matrices \code{ws}, \code{MMs}, \code{Fs}, \code{fs}, \code{N}, \code{B} and \code{F} are all of
equal size, and have one column for each age class, and one row
for each time step through the year.  So if a daily time step is
adopted, each matrix will have 366 rows with the first
representing the midnight on the last day of the previous year,
and the last representing midnight on the last day of the current
year.

If \code{yield=0}, then yield is not calculated.  If \code{yield=1}, then
only the final yield from each age class is calculated and \code{Y} is
a vector with one element for each age class.  If \code{yield=2} then
\code{Y} is the yield per time step and age class and is a matrix the
same size as \code{N} and \code{B}.
}
\examples{
## Daily time steps and 7 age classes
nsteps <- 365
Ages <- 2:8
Days <- seq(0,1,length=nsteps+1)
h <- 1/nsteps

## Ages
ages <- outer(Days,Ages,FUN="+")
## Age-length and length-weight conversions
ls <- vonBertalanffyAL(ages,t0=0.0667,K=0.5,Linf=500)
ws <- powerLW(ls,9E-10,3.32)

## Constant intra-annual natural mortality
ms <- matrix(1,nsteps+1,length(Ages))
ms <- ms/mean(trapz(ms,h))
Ms <- ctrapz(ms,h)
Msf <- final(Ms)
M <- 0.1

## Normalized within year distribution of fishing effort
fwy <- double(nsteps+1)
fwy[81:220] <- 1
fwy <- fwy/trapz(fwy,h)
## Length based selectivity
ss <- rampOgive(ls,420,30)

## Intra-annual fishing mortality - effort by selectivity
fs <- fwy*ss
Fs <- ctrapz(fs,h)
Fsf <- final(Fs)
F <- 0.2

## Deterministic initial age structure
N0 <- ageStructureD(M*Msf,F*Fsf)
N0

## Annual projection from N0
pr <- projectC(ws,M*Ms,Fs,fs,Catch=0.2811238,Nref=N0,yield=2)

## Fishing mortality and total annual yield
pr$F
sum(pr$Y[nsteps+1,])

## Plot initial abundance, projected abundance, biomass and yield
opar <- par(mfrow=c(2,2),mar=c(5,4,2,2)+0.1)
pal <- hcl(h=seq(15,375,length=8)[1:7],l=65,c=100)
plot(Ages,N0,xlab="Age",ylab="Abundance")
matplot(Days,pr$N,xlab="Day",ylab="Abundance",type="l",lty=1,col=pal)
matplot(Days,pr$B,xlab="Day",ylab="Biomass",type="l",lty=1,col=pal)
matplot(Days,pr$Y,xlab="Day",ylab="Yield",type="l",lty=1,col=pal)
par(opar)

## Annual projection given relative abundance and total
## biomass estimated from a five day mid-year survey (days 150 to 154)
Nsurvey <- c(1.0,0.67,0.46,0.33,0.25,0.18,0.13) 
Bsurvey <- 15000
pr <- projectC(ws,M*Ms,Fs,fs,Catch=1954.908,Nref=Nsurvey,nref=151:155,Bref=Bsurvey,yield=2)

## Fishing mortality and total annual yield
pr$F
sum(pr$Y[nsteps+1,])

## Plot initial abundance, projected abundance, biomass and yield
opar <- par(mfrow=c(2,2),mar=c(5,4,2,2)+0.1)
pal <- hcl(h=seq(15,375,length=8)[1:7],l=65,c=100)
plot(Ages,pr$N[1,],xlab="Age",ylab="Abundance")
matplot(Days,pr$N,xlab="Day",ylab="Abundance",type="l",lty=1,col=pal)
matplot(Days,pr$B,xlab="Day",ylab="Biomass",type="l",lty=1,col=pal)
matplot(Days,pr$Y,xlab="Day",ylab="Yield",type="l",lty=1,col=pal)
par(opar)

## Unattainable target catch - F is capped at Fmax
pr <- projectC(ws,M*Ms,Fs,fs,Catch=2,Nref=N0,yield=2,Fmax=2.5)

## Yield does not match target
sum(pr$Y[nsteps+1,])
pr$F==2.5

## Ten year projection from virgin stock collating final annual
## abundance, biomass and yield
Years <- 1:10
Nf <- Bf <- Yf <- matrix(NA,length(Years),length(Ages))

## Deterministic initial age structure assuming no fishing and 8000
## new recruits each year
N0 <- ageStructureD(M*Msf,R=8000)

## Project forward with constant catch
for(k in 1:nrow(Nf)) {
  pr <- projectC(ws,M*Ms,Fs,fs,Catch=4000,Nref=N0,yield=1)
  N0 <- advance(pr$N,R=8000)
  Nf[k,] <- final(pr$N)
  Bf[k,] <- final(pr$B)
  Yf[k,] <- pr$Y
}

## Plot annual abundance, biomass and yield
opar <- par(mfrow=c(2,2),mar=c(5,4,2,2)+0.1)
pal <- hcl(h=seq(15,375,length=8)[1:7],l=65,c=100)
matplot(Years,Nf,xlab="Year",ylab="Abundance",type="l",lty=1,col=pal)
matplot(Years,Bf,xlab="Year",ylab="Biomass",type="l",lty=1,col=pal)
matplot(Years,Yf,xlab="Year",ylab="Yield",type="l",lty=1,col=pal)
par(opar)
}
\seealso{
\code{\link[=project]{project()}}, \code{\link[=advance]{advance()}} for annual projection,
\code{\link[=ageStructureD]{ageStructureD()}}, \code{\link[=ageStructureS]{ageStructureS()}} for initial age structure
}
