% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{projectMF}
\alias{projectMF}
\title{Yearly projections for known fishing mortality (multifleet).}
\usage{
projectMF(
  ws,
  MMs,
  FFs = 0,
  Ffs = 0,
  Nref = 1,
  nref = 1,
  Bref = NA,
  bref = nref,
  yield = 2
)
}
\arguments{
\item{ws}{matrix of weight at age}

\item{MMs}{matrix of \emph{scaled} integrated natural mortalities}

\item{FFs}{matrix of \emph{scaled} integrated fishing mortalities
summed over fleets}

\item{Ffs}{array of \emph{scaled} fishing mortalities per fleet}

\item{Nref}{vector of reference abundances for each age class}

\item{nref}{vector of time steps at which the reference abundance
is estimated}

\item{Bref}{optional reference biomass}

\item{bref}{vector of time steps at which the reference biomass is
estimated}

\item{yield}{should yield be calculated for the final time step,
ever time step, or not at all (see details)}
}
\value{
\code{project} returns a list with elements * \code{N} matrix of
abundance * \code{B} matrix of biomass * \code{Y} matrix or vector of
yield
}
\description{
Project the abundance, biomass and yield in each age class forward
over one year for multiple fleets when natural and fishing
mortalities are known.
}
\details{
This is a multiple fleet variant of the \code{project} function. The
arguments are the same as for the \code{project} function except that
fishing mortalities must be specified for each fleet.
}
\examples{
## Daily time steps and 7 age classes
nsteps <- 365
Ages <- 2:8
Days <- seq(from=0, to=1, length=nsteps+1)
h <- 1/nsteps

## Ages
ages <- outer(X=Days, Y=Ages, FUN="+")
## Age-length and length-weight conversions
ls <- vonBertalanffyAL(A=ages, t0=0.0667, K=0.5, Linf=500)
ws <- powerLW(L=ls, a=9E-10, b=3.32)

## Constant intra-annual natural mortality
ms <- matrix(data=1, nrow=nsteps+1, ncol=length(x=Ages))
ms <- ms/mean(x=trapz(fs=ms, h=h))
Ms <- ctrapz(fs=ms, h=h)
Msf <- final(P=Ms)
M <- 0.1

## Normalized within year distribution of fishing effort
fwy <- double(length=nsteps+1)
fwy[81:220] <- 1
fwy <- fwy/trapz(fs=fwy, h=h)
fwy1 <- fwy2 <- fwy
## Fleet 1 doesn't fish from day 150 to 219
fwy1[151:220] <- 0
## Fleet 2 doesn't fish from day 80 to 149
fwy2[81:150] <- 0

## Length based selectivity
ss <- rampOgive(x=ls, x50=420, xrange=30)

## Intra-annual fishing mortality - effort by selectivity
fs <- array(0,c(nsteps+1,length(Ages),2))
fs[,,1] <- fwy1*ss
fs[,,2] <- fwy2*ss
Fs <- ctrapz(fs=fs, h=h)
Fsf <- Fs[nrow(Fs),,] #final(P=Fs)
F <- c(0.2,0.2)

## Deterministic initial age structure
N0 <- ageStructureD(MMsf=M*Msf, FFsf=Fsf\%*\%F)
N0

## Scale by F
FFs <- scale3(Fs,F,sum=TRUE)
Ffs <- scale3(fs,F,sum=FALSE)
## Annual projection from N0
pr <- projectMF(ws=ws, MMs=M*Ms, FFs=FFs, Ffs=Ffs, Nref=N0, yield=1)
pr$Y

colSums(pr$Y)
}
\seealso{
\code{\link[=project]{project()}}, \code{\link[=projectC]{projectC()}}, \code{\link[=advance]{advance()}} for annual
projection, \code{\link[=ageStructureD]{ageStructureD()}}, \code{\link[=ageStructureS]{ageStructureS()}} for initial age
structure
}
