% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{prRecruitsGYM}
\alias{prRecruitsGYM}
\alias{prRecruitsQuantileGYM}
\alias{prRecruitParsGYM}
\title{Proportional recruitment (GYM)}
\usage{
prRecruitsGYM(n, ps, mnA = 1)

prRecruitsQuantileGYM(q, ps, mnA = 1)

prRecruitParsGYM(
  Msf,
  mnR,
  vrR,
  r = 1,
  p = 0,
  plus = 0,
  Mbrak = seq(0, 2.5, 0.1),
  tol = 1e-08
)
}
\arguments{
\item{n}{number of recruitment events}

\item{ps}{a list of parameters returned by \code{prRecruitParGYM}}

\item{mnA}{mean number of recruits}

\item{q}{vector of quantiles}

\item{Msf}{matrix of unscaled final integrated natural mortalities}

\item{mnR}{the mean of the proportion R}

\item{vrR}{the variance of the proportion R}

\item{r}{the reference age class}

\item{p}{(ignored)}

\item{plus}{the number of additional age classes in the plus class
(\code{plus=0} implies no plus class).}

\item{Mbrak}{the values of natural mortality to assess when
bracketing a root for the nonlinear solver}

\item{tol}{the tolerance on mean R.}
}
\value{
The function \code{prRecruitParGYM} returns a list with
elements
\itemize{
\item \code{a}, \code{b} parameters of the Beta distribution
\item \code{s0} summed survivorship (T in de la Mare (94))
\item \code{B} bias correction
\item \code{M} the estimated scaling of natural mortality
}

The function \code{prRecruitsQuantileGYM} returns a vector of
quantiles from the distribution of recruit numbers.

The function \code{prRecruitsGYM} returns a vector of
random numbers of recruits.
}
\description{
Generate random recruit numbers according to the proportional
recruitment model as implemented in the GYM.
}
\details{
These functions reproduce the proportional recruits model as
implemented in the GYM and described in de la Mare (1994).  The
implementation in the GYM differs slightly from that described in de
la Mare (1994) in that a different bias correction is used.

The proportional recruit model defines R, the number of
individuals in age class r as a proportion of the population of
the same age or older. Given estimates of the mean and variance of
R from independent surveys, the model determines the scaling of
natural mortality \code{M} and random recruitment numbers that will
reproduce the observed mean and variance of R.  The model
generates random recruitment numbers as a stochastic adds of the
expected size of the population above recruitment age and the GYM
assumes the corresponding proportion is Beta distributed.

The \code{prRecruitParGYM} function generates a list of parameters
required by \code{prRecruitsGYM}.
}
\examples{
## Daily time steps and 7 age classes
nsteps <- 365
Ages <- 2:8
Days <- seq(0,1,length=nsteps+1)
h <- 1/nsteps

## Ages
ages <- outer(Days,Ages,FUN="+")
## Age-length and length-weight conversions
ls <- vonBertalanffyAL(ages,t0=0.0667,K=0.5,Linf=500)
ws <- powerLW(ls,9E-10,3.32)

## Constant intra-annual natural mortality
ms <- matrix(1,nsteps+1,length(Ages))
ms <- ms/mean(trapz(ms,h))
Ms <- ctrapz(ms,h)
Msf <- final(Ms)
M <- 0.2

## Compute model parameters that give mean R = 0.3 and var R = 0.01,
## when R is the proportion of recruits in the population
ps <- prRecruitParsGYM(mnR=0.3,vrR=0.01,Msf)
## Estimated natural mortality scaling
ps$M

## Median number of recruits
prRecruitsQuantileGYM(0.5,ps,mnA=100)

## Ten random recruit numbers
prRecruitsGYM(10,ps,mnA=100)

## Mean number of recruits
mean(prRecruitsGYM(10000,ps,mnA=100))

## Simulate age profiles and calculate R
N <- matrix(0,10000,ncol(Ms))
for(k in 1:nrow(N))
  N[k,] <- ageStructureS(prRecruitsGYM(10,ps,mnA=100),Msf,ps$M)
R <- N[,1]/rowSums(N)
## Estimate mean and variance of R
mean(R)
var(R)

## Histogram of simulated recruitments
rs <- prRecruitsGYM(100000,ps,mnA=100)
hist(rs[rs < 350],seq(0,350,5),freq=FALSE,xlab="Recruits",main="GYM")

## Compute model parameters that give mean R = 0.3 and var R = 0.01,
## when R is the proportion of individual in age class 3 as a fraction
## of individuals that are that age or older
ps <- prRecruitParsGYM(mnR=0.3,vrR=0.01,Msf,r=3)
## Estimated natural mortality scaling
ps$M

## Median number of recruits
prRecruitsQuantileGYM(0.5,ps,mnA=100)

## Ten random recruit numbers
prRecruitsGYM(10,ps,mnA=100)

## Mean number of recruits
mean(prRecruitsGYM(10000,ps,mnA=100))

## Simulate age profiles and calculate R
N <- matrix(0,100000,ncol(Ms))
for(k in 1:nrow(N))
  N[k,] <- ageStructureS(prRecruitsGYM(10,ps,mnA=100),Msf,ps$M)
R <- N[,3]/rowSums(N[,3:7])
## Estimate mean and variance of R
mean(R)
var(R)
}
\references{
de la Mare, WK. (1994) "Modelling Krill Recruitment." \emph{CCAMLR Science} 1: 49-54.
}
