% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{rampOgive}
\alias{rampOgive}
\title{Ramp ogive}
\usage{
rampOgive(x, x50, xrange)
}
\arguments{
\item{x}{a matrix or vector}

\item{x50}{mid-point of ramp}

\item{xrange}{with of ramp}
}
\value{
A matrix or vector of probabilities of the same size as x
}
\description{
Ramp shaped ogive function.
}
\details{
Compute a ramp shaped ogive function, parameterized so that \code{x50}
is the middle of the ramp, and \code{xrange} its width.
}
\examples{
x <- seq(150,450,10)
p <- rampOgive(x,300,80)
plot(x,p,type="l")
}
