% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{resampleRGYM}
\alias{resampleRGYM}
\title{Resample R}
\usage{
resampleRGYM(mnR, vrR, n)
}
\arguments{
\item{mnR}{the mean of the proportion R}

\item{vrR}{the variance of the proportion R}

\item{n}{number of independent surveys used in the estimation of
the mean and variance of R}
}
\value{
A list with elements
\itemize{
\item \code{mnR} the resampled mean R
\item \code{vrR} the resampled variance of R
}
}
\description{
Resample mean and variance of R
}
\details{
This is the resampling procedure for the mean and variance of R
implemented in the GYM. This procedure is extremely inadvisable
and should never be used.
}
\examples{
## Simulate the varianceof R
vs <- replicate(n=10000, expr=resampleRGYM(mnR=0.5, vrR=0.17, n=17)$vrR)
## The simulated variance is often larger than it is possible for the
## variance of a proprtion to be
hist(x=vs, breaks=50, xlab="Variance R", main="Resampled Variances")
}
