% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{scale3}
\alias{scale3}
\title{Scaling a 3D array}
\usage{
scale3(A, b, sum = FALSE)
}
\arguments{
\item{A}{a 3D array}

\item{b}{vector of scaling factors}

\item{sum}{sum result along third dimension}
}
\value{
the scaled array.
}
\description{
Scale a 3D array along its third dimension.
}
\details{
A convenience function that scales a 3D array along its third
dimension, optionally summing along the third dimension.
}
