% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{spawningB0D}
\alias{spawningB0D}
\title{Deterministic Spawning Biomass}
\usage{
spawningB0D(gs, ws, MMs, spawn, R = 1, plus = 0)
}
\arguments{
\item{gs}{a matrix of maturity at age}

\item{ws}{matrix of weight at age}

\item{MMs}{vector of \emph{unscaled} integrated natural mortality}

\item{spawn}{the increments in which spawning occurs.}

\item{R}{the mean numbers of recruits}

\item{plus}{the number of additional age classes in the plus class
(\code{plus=0} implies no plus class).}
}
\value{
The estimated spawning stock biomass
}
\description{
Deterministically estimate virgin spawning stock biomass.
}
\details{
Estimate virgin spawning stock biomass by deterministically
computing an initial age structure and projecting forward in time
to compute spawning stock biomass.
}
\examples{
## Daily time steps and 7 age classes
nsteps <- 365
Ages <- 2:8
Days <- seq(from=0, to=1, length=nsteps+1)
h <- 1/nsteps

## Ages
ages <- outer(X=Days, Y=Ages, FUN="+")
## Age-length and length-weight conversions
ls <- vonBertalanffyAL(A=ages, t0=0.0667, K=0.5, Linf=500)
ws <- powerLW(L=ls, a=9E-10, b=3.32)

## Constant intra-annual natural mortality
ms <- matrix(data=1, nrow=nsteps+1, ncol=length(x=Ages))
ms <- ms/mean(x=trapz(fs=ms, h=h))
Ms <- ctrapz(fs=ms, h=h)
Msf <- final(P=Ms)
M <- 0.128707

## Maturity at age
gs <- rampOgive(x=ls, x50=410, xrange=40)

## Increments in the spawning period
spawn <- 201:281

## Virgin spawning stock abundance
spawningB0D(gs=gs, ws=ws, MMs=M*Ms, spawn=spawn)

}
