% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{spawningB0S}
\alias{spawningB0S}
\title{Stochastic Median Spawning Biomass}
\usage{
spawningB0S(R, gs, ws, Ms, M, spawn, plus = FALSE)
}
\arguments{
\item{R}{a matrix of yearly recruitment numbers}

\item{gs}{a matrix of maturity at age}

\item{ws}{matrix of weight at age}

\item{Ms}{vector of \emph{unscaled} integrated natural mortality}

\item{M}{matrix of annual natural mortalities}

\item{spawn}{the increments in which spawning occurs.}

\item{plus}{whether the final class is a "plus" class}
}
\value{
The median, mean and standard deviation of the spawning
stock biomass estimated through simulation, and the number of
simulations used.
}
\description{
Stochastically estimate virgin spawning stock biomass.
}
\details{
Estimate virgin spawning stock biomass by repeatedly simulating an
initial age structure and projecting forward in time to compute
spawning stock biomass.

The matrix of random numbers of recruits \code{R} is used to simulate
initial age structures.  One spawning biomass is simulated for
every row of \code{R}, and \code{R} must have at least as many columns as
there are age classes with non neglible numbers (including age
classes aggregated into the plus group).

The annual scaling of natural mortality may vary inter-annually,
in which case \code{M} should be a matrix of the same size as \code{R}.
}
\examples{
## Daily time steps and 7 age classes
nsteps <- 365
Ages <- 2:8
Days <- seq(0,1,length=nsteps+1)
h <- 1/nsteps

## Ages
ages <- outer(Days,Ages,FUN="+")
## Age-length and length-weight conversions
ls <- vonBertalanffyAL(ages,t0=0.0667,K=0.5,Linf=500)
ws <- powerLW(ls,9E-10,3.32)

## Constant intra-annual natural mortality
ms <- matrix(1,nsteps+1,length(Ages))
ms <- ms/mean(trapz(ms,h))
Ms <- ctrapz(ms,h)
Msf <- final(Ms)

## Compute model parameters that give mean R = 0.3 and var R = 0.01,
## when R is the proportion of recruits in the population
ps <- prRecruitParsIB(mnR=0.2,vrR=0.01,Msf)
## Estimated natural mortality scaling
ps$M

R <- matrix(prRecruitsIB(10000,ps),1000,10)

## Maturity at age
gs <- rampOgive(ls,410,40)

## Increments in the spawning period
spawn <- 201:281

## Virgin spawning stock abundance
spawningB0S(R,ws,gs,Ms,ps$M,spawn)
}
