% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{initial}
\alias{initial}
\alias{final}
\alias{meanStock}
\alias{spawningStock}
\alias{exploitableStock}
\alias{vulnerableStock}
\title{Stock summaries}
\usage{
initial(P)

final(P)

meanStock(P, ws = 1, period)

spawningStock(P, gs, period)

exploitableStock(P, ss, period)

vulnerableStock(P, ss, fwy, period)
}
\arguments{
\item{P}{projected abundance or biomass}

\item{ws}{matrix of weights}

\item{period}{the time steps corresponding to the
spawning/monitoring period}

\item{gs}{matrix of fraction mature at age}

\item{ss}{matrix of selectivity at age}

\item{fwy}{vector of within year distribution of effort for the
year.}
}
\value{
\code{initial} and \code{final} return the first and last rows the yearly
projected abundance or biomass.

\code{meanStock} returns a weighted mean abundance or biomass in a
given period.

\code{spawningStock} returns the mean abundance or biomass of mature
fish in a given time period.

\code{exploitableStock} returns the mean abundance or biomass that
could be exploited by fishing in a given time period.

\code{vulnerableStock} returns the effort weighted mean exploitable
abundance or biomass in a given time period.
}
\description{
Calculate the stock summaries for a given monitroing period.
}
\details{
The \code{spawningStock} and \code{exploitableStock} functions are just
special cases of \code{meanStock}.  \code{vulnerableStock} differs in that
it also weights by the within year fishing effort.
}
\examples{
## Daily time steps and 7 age classes
nsteps <- 365
Ages <- 2:8
Days <- seq(from=0, to=1, length=nsteps+1)
h <- 1/nsteps

## Ages
ages <- outer(X=Days, Y=Ages, FUN="+")
## Age-length and length-weight conversions
ls <- vonBertalanffyAL(A=ages, t0=0.0667, K=0.5, Linf=500)
ws <- powerLW(L=ls, a=9E-10, b=3.32)

## Constant intra-annual natural mortality
ms <- matrix(data=1, nrow=nsteps+1, ncol=length(x=Ages))
ms <- ms/mean(x=trapz(fs=ms, h=h))
Ms <- ctrapz(fs=ms, h=h)
Msf <- final(P=Ms)
M <- 0.1

## Normalized within year distribution of fishing effort
fwy <- double(length=nsteps+1)
fwy[81:220] <- 1
fwy <- fwy/trapz(fs=fwy, h=h)
## Length based selectivity
ss <- rampOgive(x=ls, x50=420, xrange=30)

## Intra-annual fishing mortality - effort by selectivity
fs <- fwy*ss
Fs <- ctrapz(fs=fs, h=h)
Fsf <- final(P=Fs)
F <- 0.2

## Deterministic initial age structure
N0 <- ageStructureD(MMsf=M*Msf, FFsf=F*Fsf)
N0

## Annual projection from N0
pr <- project(ws=ws, MMs=M*Ms, FFs=F*Fs, Ffs=F*fs, Nref=N0, yield=2)

## Initial abundance
initial(P=pr$N)
## Final abundance
final(P=pr$N)

## Maturity at age
gs <- rampOgive(x=ls, x50=410, xrange=40)

## Increments in the spawning period
spawn <- 201:281

## Spawning stock abundance
spawningStock(P=pr$N, gs=gs, period=spawn)
## Spawning stock biomass
spawningStock(P=pr$B, gs=gs, period=spawn)

## Monitoring period within fishing season
monitor <- 101:151
## Exploitable biomass
exploitableStock(P=pr$B, ss=ss, period=monitor)
## Weighted vulnerable biomass
vulnerableStock(P=pr$B, ss=ss, fwy=fwy, period=monitor)

## Monitoring period spans fishing season
monitor <- 41:101
## Exploitable biomass
exploitableStock(P=pr$B, ss=ss, period=monitor)
## Vulnerable biomass weights by effort
vulnerableStock(P=pr$B, ss=ss, fwy=fwy, period=monitor)

## No effort in monitoring period
monitor <- 11:31
## Exploitable biomass
exploitableStock(P=pr$B, ss=ss, period=monitor)
## Vulnerable biomass weights by effort
vulnerableStock(P=pr$B, ss=ss, fwy=fwy, period=monitor)
}
