% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{surveySurvival}
\alias{surveySurvival}
\title{Survival to a survey period.}
\usage{
surveySurvival(yr, cls, s1, s2, Ms, M, Fs = 0, F = 0, rcls = 1)
}
\arguments{
\item{yr}{vector of survey year}

\item{cls}{vector of survey age class}

\item{s1}{vector of the first time step in the survey}

\item{s2}{vector of the final time step in the survey}

\item{Ms}{matrix of \emph{unscaled} integrated natural mortality}

\item{M}{vector of annual natural mortalities}

\item{Fs}{matrix of \emph{unscaled} integrated fishing mortality}

\item{F}{vector of annual fishing mortalities}

\item{rcls}{the reference age class to adjust to}
}
\value{
Returns a vector of the mean survival from time of
recruitment to the survey period.
}
\description{
Compute the scalings required to adjust surveyed age class
abundances to initial abundances at a reference age.
}
\details{
Given the age class, the year and the time steps within the year
at which the age class was surveyed, this function computes the
total survival from the start of the year in which the cohort were
in the reference age class to the survey period.  If the surveyed
age class is younger than the reference class, the reciporical of
the total survival from the survey period to the start of the year
that the cohort will be in the reference class is computed.

If there is inter-annual variability in natural or fishing
mortality, the survey years must be labelled so that \code{yr==1}
corresponds to the first element of the vector of \code{M} and/or \code{F},
and it is no possible to compute the survival for cohorts that
recruit before year 1.

If there is no inter-annual variablity in natural or fishing
mortality, the survey year is irrelevant.
}
\examples{
## Daily time steps and 7 age classes
nsteps <- 365
Ages <- 2:8
Days <- seq(0,1,length=nsteps+1)
h <- 1/nsteps

## Constant intra-annual natural mortality
ms <- matrix(1,nsteps+1,length(Ages))
ms <- ms/mean(trapz(ms,h))
Ms <- ctrapz(ms,h)

## Survey year, period and age classes
svy <- data.frame(yr=3:5,s1=c(190,220,150),s2=c(201,231,161))
svy <- cbind(svy[rep(1:3,each=7),],cls=1:7)
head(svy)

## Constant mortality
M <- 0.2

## Survival to the survey period from age class 1
surveySurvival(svy$yr,svy$cls,svy$s1,svy$s2,Ms,M)

## Survival to the survey period from age class 3
surveySurvival(svy$yr,svy$cls,svy$s1,svy$s2,Ms,M,rcls=3)

## Variable mortality
M <- rgamma(10,20,100)
M

## Survival cannot be projected outside the period for which mortality
## is specified.
surveySurvival(svy$yr,svy$cls,svy$s1,svy$s2,Ms,M)
}
