% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{trapzMeans}
\alias{trapzMeans}
\title{Average over a time interval.}
\usage{
trapzMeans(fs)
}
\arguments{
\item{fs}{a vector or matrix of function values}
}
\value{
Returns the approximation to the average of the function or functions
represented by \code{fs}.
}
\description{
Compute averages over a time interval, but the interval represents
a single point in time, the value at that time point is returned.
}
\details{
Given a vector \code{fs} of function values evaluated over an even
partition of an interval, \code{trapzMean} approximates the average
over the interval by the composite \href{wikipedia.org/wiki/Trapezoidal_rule}{trapezoidal rule}.

In each case, if \code{fs} does not span a time interval but represents
a single point in time, the value at that time point is returned.
}
\examples{
## The average of 2*x from 1 to 5
x <- 1:5
trapzMeans(fs=2*x)
trapz(fs=2*x, h=1)/4

## For a single point return the function value
x <- 1
trapzMeans(fs=2*x)

## Average of cos(x), sin(x) from 0 to 2 pi
h <- 2*pi/100
x <- seq(from=0, to=2*pi, by=h)
cs <- cbind(cos(x), sin(x))
trapzMeans(fs=cs)
}
