% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{vonBertalanffyAL}
\alias{vonBertalanffyAL}
\alias{vonBertalanffyLA}
\alias{vonBertalanffyRAL}
\alias{vonBertalanffyRLA}
\title{von Bertalanffy Growth}
\usage{
vonBertalanffyAL(A, t0, K, Linf)

vonBertalanffyLA(L, t0, K, Linf)

vonBertalanffyRAL(A, t0, K, Linf, f0 = 0, f1 = 1)

vonBertalanffyRLA(L, t0, K, Linf, f0 = 0, f1 = 1)
}
\arguments{
\item{A}{vector of ages}

\item{t0}{reference time parameter}

\item{K}{growth rate parameter}

\item{Linf}{asymptotic length parameter}

\item{L}{vector of lengths}

\item{f0}{fraction of year at which growth starts}

\item{f1}{fraction of year at which growth ends}
}
\value{
\code{vonBertalanffyAL} returns the vector of lengths
corresponding to the ages \code{A}, and \code{vonBertalanffy} returns a
vector of ages corresponding to the lengths \code{L}.
}
\description{
The von Bertalanffy growth curve and its inverse.
}
\details{
These functions compute the von Bertalanffy growth curve using the
standard parameterization implemented in the GYM.

The functions \code{vonBertalanffyAL} and \code{vonBertalanffyLA} use a von
Bertalanffy growth to map age to length and length to age. The
variants \code{vonBertalanffyRAL} and \code{vonBertalanffyRLA} assume that
growth is restricted to some fraction of each year.
}
\examples{
age <- 2:8

## Convert age to length
len <- vonBertalanffyAL(A=age, t0=0.0667, K=0.5, Linf=500)
len

## Convert length to age
age <- vonBertalanffyLA(L=len, t0=0.0667, K=0.5, Linf=500)
age

## Daily time steps and 7 age classes
nsteps <- 365
Ages <- 2:8
Days <- seq(from=0, to=1, length=nsteps+1)
h <- 1/nsteps

## Ages
ages <- outer(X=Days, Y=Ages, FUN="+")

## Age-length conversion
ls <- vonBertalanffyAL(A=ages, t0=0.0667, K=0.5, Linf=500)
matplot(x=Days, y=ls, type="l", lty=1, xlab="Day", ylab="Length", main="Length at Age")

## Age-length conversion - growth occurs in middle of the year
ls <- vonBertalanffyRAL(A=ages, t0=0.0667, K=0.5, Linf=500, f0=0.3, f1=0.7)
matplot(x=Days, y=ls, type="l", lty=1, xlab="Day", ylab="Length", main="Length at Age")
}
\seealso{
\code{\link[=powerLW]{powerLW()}} for length-weight modelling.
}
