% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{logisticOgive}
\alias{logisticOgive}
\title{Ramp ogive}
\usage{
logisticOgive(x, x50, x95)
}
\arguments{
\item{x}{a matrix or vector}

\item{x50}{50th percentile}

\item{x95}{95th percentile}
}
\value{
A matrix or vector of probabilities of the same size as x
}
\description{
Logistic shaped ogive function.
}
\details{
Compute a logistic shaped ogive function, parameterized so that \code{x50}
is the middle of the ramp, and \code{x95} is the 95th percentile.
}
\examples{
x <- seq(from=150, to=450, by=10)
p <- logisticOgive(x, x50=300, x95=380)
plot(x=x, y=p, type="l")
}
