% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grym.R
\name{spawningB0D}
\alias{spawningB0D}
\title{Deterministic Spawning Biomass}
\usage{
spawningB0D(gs, ws, MMs, spawn, R = 1, plus = 0)
}
\arguments{
\item{gs}{a matrix of maturity at age}

\item{ws}{matrix of weight at age}

\item{MMs}{vector of \emph{unscaled} integrated natural mortality}

\item{spawn}{the increments in which spawning occurs.}

\item{R}{the mean numbers of recruits}

\item{plus}{the number of additional age classes in the plus class
(\code{plus=0} implies no plus class).}
}
\value{
The estimated spawning stock biomass
}
\description{
Deterministically estimate virgin spawning stock biomass.
}
\details{
Estimate virgin spawning stock biomass by deterministically
computing an initial age structure and projecting forward in time
to compute spawning stock biomass.
}
\examples{
## Daily time steps and 7 age classes
nsteps <- 365
Ages <- 2:8
Days <- seq(0,1,length=nsteps+1)
h <- 1/nsteps

## Ages
ages <- outer(Days,Ages,FUN="+")
## Age-length and length-weight conversions
ls <- vonBertalanffyAL(ages,t0=0.0667,K=0.5,Linf=500)
ws <- powerLW(ls,9E-10,3.32)

## Constant intra-annual natural mortality
ms <- matrix(1,nsteps+1,length(Ages))
ms <- ms/mean(trapz(ms,h))
Ms <- ctrapz(ms,h)
Msf <- final(Ms)
M <- 0.128707

## Maturity at age
gs <- rampOgive(ls,410,40)

## Increments in the spawning period
spawn <- 201:281

## Virgin spawning stock abundance
spawningB0D(ws,gs,M*Ms,spawn)
}
